+++ examples/js/pxlNavLoader_switchSpace.js
@@ -1,4 +1,4 @@
-// pxlNav v0.0.18 -  Javascript Launcher
+// pxlNav v0.0.28 -  Javascript Launcher
 //  Written by Kevin Edzenga; 2024,2025
 //
 // -- -- -- -- -- -- -- -- -- -- -- -- -- --
@@ -6,11 +6,10 @@
 //   This is an example implementation of `pxlNav` in a project;
 //     Tieing in `ProcPages` to manage the pages of the site,
 //       Listening to / triggering events on `pxlNav`
-//   For `pxlNav` scripting, the entry-point is `./Source/js/pxlNavCore.js`
+//   For `pxlNav` scripting, the entry-point is `./src/js/pxlNav.js`
 //
 
 import { pxlNav, pxlNavVersion, pxlEnums, pxlUserSettings, pxlOptions } from './pxlNav.esm.js';
-
 
 // Console logging level
 //   Options are - NONE, ERROR, WARN, INFO
@@ -18,7 +17,7 @@ const verbose = pxlEnums.VERBOSE_LEVEL.INFO;
 
 // The Title of your Project
 //   This will be displayed on the load bar
-const projectTitle = "pxlNav : The Outlet";
+const projectTitle = "pxlNav :: The Outlet";
 
 // pxlRoom folder path, available to change folder names or locations if desired
 const pxlRoomRootPath = "../js/pxlRooms";
@@ -32,6 +31,21 @@ const showOnboarding = true;
 // Current possible rooms - "OutletEnvironment", "VoidEnvironment"
 const bootRoomList = ["OutletEnvironment","VoidEnvironment"];
 const startingRoom = bootRoomList[0];
+
+// -- -- --
+
+// Optional : Display a Frame Rate (FPS) counter in a console
+//              This is an example callback to show the FPS in a div
+// You can also enable this with `?showfps=1` in the URL
+//   If you leave the `uriSearch` code below vvv
+
+// Name of the div to display verbose messages
+//   This uses callback events from `pxlNav` to display messages
+const verboseDivName = 'verbErrorConsole';
+
+// Display FPS in the `verboseDivName` div
+//   Useful for debugging and performance testing
+const enableFPSDisplay = false;
 
 // -- -- --
 
@@ -58,6 +72,7 @@ userSettings['height']['stepSize'] = 5; // Max step height in units
 userSettings['movement']['scalar'] = 1.0; // Overall movement rate scalar
 userSettings['movement']['max'] = 10.0; // Max movement speed
 userSettings['movement']['easing'] = 0.55; // Easing rate between Step() calls
+userSettings['look']['mobile']['invert'] = true; // Invert the look controls on mobile; Default is `True`
 userSettings['headBounce']['height'] = 0.3; // Bounce magnitude in units
 userSettings['headBounce']['rate'] = 0.025; // Bounce rate per Step()
 userSettings['headBounce']['easeIn'] = 0.03; // When move key is pressed, the ease into bounce; `bounce * ( boundInf + easeIn )`
@@ -65,8 +80,8 @@ userSettings['headBounce']['easeOut'] = 0.95; // When move key is let go, the ea
 userSettings['jump']['impulse'] = 0.75; // Jump impulse force applied to the player while holding the jump button
 userSettings['jump']['holdMax'] = 2.85; // Max influence of holding the jump button on current jump; in seconds
 userSettings['jump']['repeatDelay'] = 0.08; // Delay between jumps when holding the jump button
-userSettings['gravity']['UPS'] = 0.3; // Units per Step() per Step()
-userSettings['gravity']['Max'] = 15.5; // Max gravity rate
+userSettings['gravity']['ups'] = 0.3; // Units per Step() per Step()
+userSettings['gravity']['max'] = 15.5; // Max gravity rate
 
 // -- -- --
 
@@ -77,12 +92,12 @@ const targetFPS = {
   'mobile' : 30
 };
 
-// Render Overscan - Render larger than the screen resolution to reduce edge artifacts
-//   Since mobile devices have a lower resolution, overscan may help
-// Default is - PC = 0.0  -&-  Mobile = 0.0
-const renderOverscan = {
-  'pc' : 0.0,
-  'mobile' : 0.5
+// Render Resolution Scale
+//   Since mobile devices have a lower resolution, up scaling may help
+// Default is - PC = 1.0  -&-  Mobile = 1.0
+const renderScale = {
+  'pc' : 1.0,
+  'mobile' : 1.5
 }
 
 // Anti-aliasing level
@@ -119,6 +134,42 @@ const collisionScale = {
 };
 
 
+
+// -- -- -- -- -- -- -- -- -- -- -- -- -- -- //
+// -- -- -- -- -- -- -- -- -- -- -- -- -- -- //
+// -- -- -- -- -- -- -- -- -- -- -- -- -- -- //
+
+
+// Add some search uri parameters to the page
+//   These are used to set the target FPS and renderScale
+// This is not needed for pxlNav, but is useful for testing
+//
+// Find search parameters in the URL for procstack.github.io
+//   Not needed for pxlNav
+// Note : procPages clears the search parameters on the page
+//          So search is lost on page change,
+//            Running before procPages.init() is needed
+let uriSearch = window.location.search;
+
+// Check hash for fps and renderScale
+let searchParams = new URLSearchParams(uriSearch);
+let showFPS = searchParams.has('showfps') ? !!parseInt(searchParams.get('showfps')) : false;
+if( searchParams.has('fps') ){
+  let fps = parseInt(searchParams.get('fps'));
+  if( fps > 0 ){
+    targetFPS.pc = fps;
+    targetFPS.mobile = fps;
+  }
+}
+if( searchParams.has('scale') ){
+  let scale = parseFloat(searchParams.get('scale'));
+  if( scale > 0 ){
+    renderScale.pc = scale;
+    renderScale.mobile = scale;
+  }
+}
+
+
 // -- -- -- -- -- -- -- -- -- -- -- -- -- -- //
 // -- -- -- -- -- -- -- -- -- -- -- -- -- -- //
 // -- -- -- -- -- -- -- -- -- -- -- -- -- -- //
@@ -133,7 +184,7 @@ const collisionScale = {
 let pxlNavOptions = Object.assign({},pxlOptions);
 pxlNavOptions.verbose = verbose;
 pxlNavOptions.fps = targetFPS;
-pxlNavOptions.overscan = renderOverscan;
+pxlNavOptions.renderScale = renderScale;
 pxlNavOptions.userSettings = userSettings;
 pxlNavOptions.antiAliasing = antiAliasing;
 pxlNavOptions.collisionScale = collisionScale;
@@ -150,6 +201,20 @@ pxlNavOptions.loaderPhrases = loaderPhrases;
 
 // Create the pxlNav environment manager
 const pxlNavEnv = new pxlNav( pxlNavOptions, projectTitle, startingRoom, bootRoomList );
+
+
+// -- -- --
+
+// Subscribe to events emitted from pxlNav for callback handling
+//   Non loop - pxlNavObj.subscribe("pxlNavEventNameHere", procPages.functionName.bind(procPages));
+
+/* Uncomment to add custom event handling */
+/*
+const pageListenEvents = [ "booted", "shaderEditorVis", "roomChange-End", "fromRoom" ];
+pageListenEvents.forEach( (e)=>{
+  pxlNavEnv.subscribe( e, procPages.eventHandler.bind(procPages) );
+});
+*/
 
 
 // -- -- --
@@ -180,6 +245,46 @@ if( switchButton && switchButton.hasAttribute("roomToggles") ){
       pxlNavEnv.trigger( 'warptoroom', nextVal );
     }
   });
+}
+
+
+// -- -- --
+
+// Display the frame rate (FPS) in the `verboseDivName` div
+//   This is useful for debugging and performance testing
+// Listen to the `render-prep` call, occuring before the render occurs per frame
+//   Then have it update with a 4-frame average of the FPS
+// Display the frame rate (FPS) in the `verboseDivName` div
+function setupFPSDisplay( pxlNavEnv, verboseDivName, avgCount = 4 ){
+  let verboseConsole = document.getElementById(verboseDivName);
+  if( !verboseConsole ){
+    if( verbose >= pxlEnums.VERBOSE_LEVEL.WARN ){
+      console.warn("No verbose console found with id: " + verboseDivName);
+    }
+    return;
+  }
+
+  let skipRunner = 0;
+  let avgFPS = 0;
+  let prevTime = 0;
+
+  pxlNavEnv.subscribe('render-prep', (e) => {
+    skipRunner++;
+    let delta = (1 / (e.value.time - prevTime));
+    prevTime = e.value.time;
+    avgFPS += delta;
+    if (skipRunner >= avgCount) {
+      avgFPS = (avgFPS / skipRunner).toFixed(2);
+      verboseConsole.innerText = avgFPS;
+      avgFPS = 0;
+      skipRunner = 0;
+    }
+  });
+}
+
+// Then in the initialization code:
+if( enableFPSDisplay || showFPS ){
+  setupFPSDisplay(pxlNavEnv, verboseDivName);
 }