+++ src/pxlNavLoader_switchSpace.js
@@ -1,4 +1,4 @@
-// pxlNav v0.0.18 -  Javascript Launcher
+// pxlNav v0.0.28 -  Javascript Launcher
 //  Written by Kevin Edzenga; 2024,2025
 //
 // -- -- -- -- -- -- -- -- -- -- -- -- -- --
@@ -6,15 +6,15 @@
 //   This is an example implementation of `pxlNav` in a project;
 //     Tieing in `ProcPages` to manage the pages of the site,
 //       Listening to / triggering events on `pxlNav`
-//   For `pxlNav` scripting, the entry-point is `./Source/js/pxlNavCore.js`
+//   For `pxlNav` scripting, the entry-point is `./src/js/pxlNav.js`
 //
 
-import { pxlNav, pxlNavVersion, pxlEnums, pxlUserSettings, pxlOptions } from './pxlNavs.js';
+import { pxlNav, pxlNavVersion, pxlEnums, pxlUserSettings, pxlOptions } from './pxlNav.js';
 
 
 // Console logging level
 //   Options are - NONE, ERROR, WARN, INFO
-const verbose = pxlEnums.VERBOSE_LEVEL.INFO;
+const verbose = pxlEnums.VERBOSE_LEVEL.NONE;
 
 // The Title of your Project
 //   This will be displayed on the load bar
@@ -32,6 +32,21 @@ const showOnboarding = false;
 // Current possible rooms - "OutletEnvironment", "VoidEnvironment"
 const bootRoomList = ["CampfireEnvironment","VoidEnvironment"];
 const startingRoom = bootRoomList[0];
+
+// -- -- --
+
+// Optional : Display a Frame Rate (FPS) counter in a console
+//              This is an example callback to show the FPS in a div
+// You can also enable this with `?showfps=1` in the URL
+//   If you leave the `uriSearch` code below vvv
+
+// Name of the div to display verbose messages
+//   This uses callback events from `pxlNav` to display messages
+const verboseDivName = 'verbErrorConsole';
+
+// Display FPS in the `verboseDivName` div
+//   Useful for debugging and performance testing
+const enableFPSDisplay = false;
 
 // -- -- --
 
@@ -58,6 +73,7 @@ userSettings['height']['stepSize'] = 5; // Max step height in units
 userSettings['movement']['scalar'] = 1.0; // Overall movement rate scalar
 userSettings['movement']['max'] = 10.0; // Max movement speed
 userSettings['movement']['easing'] = 0.55; // Easing rate between Step() calls
+userSettings['look']['mobile']['invert'] = true; // Invert the look controls on mobile; Default is `True`
 userSettings['headBounce']['height'] = 0.3; // Bounce magnitude in units
 userSettings['headBounce']['rate'] = 0.025; // Bounce rate per Step()
 userSettings['headBounce']['easeIn'] = 0.03; // When move key is pressed, the ease into bounce; `bounce * ( boundInf + easeIn )`
@@ -65,8 +81,8 @@ userSettings['headBounce']['easeOut'] = 0.95; // When move key is let go, the ea
 userSettings['jump']['impulse'] = 0.75; // Jump impulse force applied to the player while holding the jump button
 userSettings['jump']['holdMax'] = 2.85; // Max influence of holding the jump button on current jump; in seconds
 userSettings['jump']['repeatDelay'] = 0.08; // Delay between jumps when holding the jump button
-userSettings['gravity']['UPS'] = 0.3; // Units per Step() per Step()
-userSettings['gravity']['Max'] = 15.5; // Max gravity rate
+userSettings['gravity']['ups'] = 0.3; // Units per Step() per Step()
+userSettings['gravity']['max'] = 15.5; // Max gravity rate
 
 // -- -- --
 
@@ -80,7 +96,7 @@ const targetFPS = {
 // Render Resolution Scale
 //   Since mobile devices have a lower resolution, up scaling may help
 // Default is - PC = 1.0  -&-  Mobile = 1.0
-pxlNavOptions.renderScale = renderScale;
+const renderScale = {
   'pc' : 1.0,
   'mobile' : 1.5
 }
@@ -98,7 +114,7 @@ const shadowMapBiasing = pxlEnums.SHADOW_MAP.SOFT;
 // Set camera to static Camera Positions
 //   Locations pulled from the 'Camera' group in the pxlRoom's FBX file
 // Default is `false`
-const enableStaticCamera = false;
+const enableStaticCamera = true;
 
 // If using static cameras, allow the user to rotate the camera
 //  Default is `false`
@@ -117,6 +133,42 @@ const collisionScale = {
   'gridSize' : 100,
   'gridReference' : 1000
 };
+
+
+
+// -- -- -- -- -- -- -- -- -- -- -- -- -- -- //
+// -- -- -- -- -- -- -- -- -- -- -- -- -- -- //
+// -- -- -- -- -- -- -- -- -- -- -- -- -- -- //
+
+
+// Add some search uri parameters to the page
+//   These are used to set the target FPS and renderScale
+// This is not needed for pxlNav, but is useful for testing
+//
+// Find search parameters in the URL for procstack.github.io
+//   Not needed for pxlNav
+// Note : procPages clears the search parameters on the page
+//          So search is lost on page change,
+//            Running before procPages.init() is needed
+let uriSearch = window.location.search;
+
+// Check hash for fps and renderScale
+let searchParams = new URLSearchParams(uriSearch);
+let showFPS = searchParams.has('showfps') ? !!parseInt(searchParams.get('showfps')) : false;
+if( searchParams.has('fps') ){
+  let fps = parseInt(searchParams.get('fps'));
+  if( fps > 0 ){
+    targetFPS.pc = fps;
+    targetFPS.mobile = fps;
+  }
+}
+if( searchParams.has('scale') ){
+  let scale = parseFloat(searchParams.get('scale'));
+  if( scale > 0 ){
+    renderScale.pc = scale;
+    renderScale.mobile = scale;
+  }
+}
 
 
 // -- -- -- -- -- -- -- -- -- -- -- -- -- -- //
@@ -154,33 +206,71 @@ const pxlNavEnv = new pxlNav( pxlNavOptions, projectTitle, startingRoom, bootRoo
 
 // -- -- --
 
-// <div id="roomToggle" roomToggles="VoidEnvironment:Void Space;OutletEnvironment:Field">Void Space</div>
+// Subscribe to events emitted from pxlNav for callback handling
+//   Non loop - pxlNavObj.subscribe("pxlNavEventNameHere", procPages.functionName.bind(procPages));
 
-let switchButton = document.getElementById("roomToggle");
-if( switchButton && switchButton.hasAttribute("roomToggles") ){
-  let roomValues = switchButton.getAttribute("roomToggles").split(";");
-  let roomLabelDict = {};
-  roomValues.forEach(curVal => {
-    let curPair = curVal.split(":");
-    roomLabelDict[curPair[0]] = curPair[1];
-  });
-  
-  switchButton.addEventListener("click", function(){
-    let switchButtonObj = document.getElementById("roomToggle");
-    if( switchButtonObj && switchButtonObj.hasAttribute("curRoom") ){ 
-      let curVal = switchButtonObj.getAttribute("curRoom");
-      let nextVal = "";
-      let labelKeys = Object.keys(roomLabelDict);
-      let curIndex = labelKeys.indexOf(curVal);
-      curIndex = (curIndex + 1) % labelKeys.length;
-      nextVal = labelKeys[curIndex];
-      switchButtonObj.innerText = roomLabelDict[nextVal];
-      switchButtonObj.setAttribute("curRoom", nextVal);
-      //console.log( "Switching to room: ", nextVal );  
-      pxlNavEnv.trigger( 'warptoroom', nextVal );
+/* Uncomment to add custom event handling */
+/*
+const pageListenEvents = [ "booted", "shaderEditorVis", "roomChange-End", "fromRoom" ];
+pageListenEvents.forEach( (e)=>{
+  pxlNavEnv.subscribe( e, procPages.eventHandler.bind(procPages) );
+});
+*/
+
+// -- -- --
+
+// Display the frame rate (FPS) in the `verboseDivName` div
+//   This is useful for debugging and performance testing
+// Listen to the `render-prep` call, occuring before the render occurs per frame
+//   Then have it update with a 4-frame average of the FPS
+// Display the frame rate (FPS) in the `verboseDivName` div
+function setupFPSDisplay( pxlNavEnv, verboseDivName, avgCount = 4 ){
+  let verboseConsole = document.getElementById(verboseDivName);
+  if( !verboseConsole ){
+    if( verbose >= pxlEnums.VERBOSE_LEVEL.WARN ){
+      console.warn("No verbose console found with id: " + verboseDivName);
+    }
+    return;
+  }
+
+  let skipRunner = 0;
+  let avgFPS = 0;
+  let prevTime = 0;
+
+  pxlNavEnv.subscribe('render-prep', (e) => {
+    skipRunner++;
+    let delta = (1 / (e.value.time - prevTime));
+    prevTime = e.value.time;
+    avgFPS += delta;
+    if (skipRunner >= avgCount) {
+      avgFPS = (avgFPS / skipRunner).toFixed(2);
+      verboseConsole.innerText = avgFPS;
+      avgFPS = 0;
+      skipRunner = 0;
     }
   });
 }
+
+// Then in the initialization code:
+if( enableFPSDisplay || showFPS ){
+  setupFPSDisplay(pxlNavEnv, verboseDivName);
+}
+
+
+// -- -- --
+
+// Function to warp to the given camera location in the starting room
+function switchToCameraLocation( cameraName ){
+  pxlNavEnv.trigger( "warpToRoom", startingRoom, cameraName );
+}
+
+// Bind the `warpToRoom` event to the `pxlNav` object once it is fully loaded
+//   Listen to the `booted` event to know when the `pxlNav` object is ready to set the camera location
+pxlNavEnv.subscribe( "booted", ( e ) => {
+  // Switch to the starting camera location
+  switchToCameraLocation( "init" );
+});
+
 
 
 // -- -- --